API

Modules

Types and constants

RadiationDetectorDSP.AbstractRadFIRFilter

RadiationDetectorDSP.AbstractRadIIRFilter

RadiationDetectorDSP.AbstractRadSigFilter

RadiationDetectorDSP.AbstractRadSigFilterInstance

RadiationDetectorDSP.AbstractSamples

RadiationDetectorDSP.ArrayOfSimilarSamples

RadiationDetectorDSP.BiquadFilter

RadiationDetectorDSP.CRFilter

RadiationDetectorDSP.ConvolutionFilter

RadiationDetectorDSP.ConvolutionMethod

RadiationDetectorDSP.DNIMethod

RadiationDetectorDSP.DifferentiatorFilter

RadiationDetectorDSP.DirectConvolution

RadiationDetectorDSP.FFTConvolution

RadiationDetectorDSP.FilteringType

RadiationDetectorDSP.FirstOrderIIR

RadiationDetectorDSP.IntegratorCRFilter

RadiationDetectorDSP.IntegratorFilter

RadiationDetectorDSP.IntegratorModCRFilter

RadiationDetectorDSP.InvCRFilter

RadiationDetectorDSP.InvModCRFilter

RadiationDetectorDSP.InvRCFilter

RadiationDetectorDSP.LinearFiltering

RadiationDetectorDSP.MaybeWithUnits

RadiationDetectorDSP.ModCRFilter

RadiationDetectorDSP.NonlinearFiltering

RadiationDetectorDSP.PolynomialDNI

RadiationDetectorDSP.RCFilter

RadiationDetectorDSP.RealQuantity

RadiationDetectorDSP.SamplesOrWaveform

RadiationDetectorDSP.SamplingInfo

RadiationDetectorDSP.SavitzkyGolayFilter

RadiationDetectorDSP.SignalEstimator

RadiationDetectorDSP.SimpleCSAFilter

RadiationDetectorDSP.TrapezoidalChargeFilter

RadiationDetectorDSP.TruncateFilter

RadiationDetectorDSP.ZACChargeFilter

Functions and macros

RadiationDetectorDSP.add_rect_pulse!

RadiationDetectorDSP.bc_rdfilt

RadiationDetectorDSP.bc_rdfilt!

RadiationDetectorDSP.charge_trapflt!

RadiationDetectorDSP.cr_filter

RadiationDetectorDSP.create_zac_filter

RadiationDetectorDSP.crmod_filter

RadiationDetectorDSP.dfilt!

RadiationDetectorDSP.differentiator_filter

RadiationDetectorDSP.elsmplinfo

RadiationDetectorDSP.flt_input_length

RadiationDetectorDSP.flt_input_smpltype

RadiationDetectorDSP.flt_output_length

RadiationDetectorDSP.flt_output_smpltype

RadiationDetectorDSP.flt_output_time_axis

RadiationDetectorDSP.fltinstance

RadiationDetectorDSP.gen_rect_pulse

RadiationDetectorDSP.integrator_cr_filter

RadiationDetectorDSP.integrator_crmod_filter

RadiationDetectorDSP.integrator_filter

RadiationDetectorDSP.inv_cr_filter

RadiationDetectorDSP.inv_crmod_filter

RadiationDetectorDSP.inv_rc_filter

RadiationDetectorDSP.rc_filter

RadiationDetectorDSP.rdfilt

RadiationDetectorDSP.shift_waveform

RadiationDetectorDSP.simple_csa_response_filter

RadiationDetectorDSP.smplinfo

RadiationDetectorDSP.zac_charge_filter_coeffs

RadiationDetectorDSP.zac_current_filter_integral

RadiationDetectorDSP.zac_current_filter_shape

Documentation

RadiationDetectorDSP.AbstractRadFIRFilter — Type

abstract type AbstractRadFIRFilter <: AbstractRadLinearFilter

Abstract type for FIR filters.

RadiationDetectorDSP.AbstractRadIIRFilter — Type

abstract type AbstractRadIIRFilter <: AbstractRadSigFilter{LinearFiltering}

Abstract type for IIR filters.

RadiationDetectorDSP.AbstractRadSigFilter — Type

abstract type AbstractRadSigFilter{FT<:FilteringType} <: Function

Abstract type for signal filters.

Filters are callable as (flt::AbstractRadSigFilter)(input) and come with specialized broadcasting.

Subtypes of AbstractRadSigFilter must implement

fltinstance(flt::AbstractRadSigFilter, si::SamplingInfo)::[`AbstractRadSigFilterInstance`](@ref)

Invertible filters should also implement

InverseFunctions.inverse(flt::SomeFilter)

Not that while a filter may have an inverse, it may (often depending on the filter paramters) be very unstable in the presence of additional noise. So the inverse may be not be useful to deconvolve signals in practical applications.

RadiationDetectorDSP.AbstractRadSigFilterInstance — Type

abstract type AbstractRadSigFilterInstance{FT<:FilteringType}

Abstract type for signal filter instances. Filter instances are specilized to a specific length and numerical type of input and output.

Filter instances are callable as (fi::SomeFilterInstance)(input) and come with specialized broadcasting.

Subtypes of AbstractRadSigFilterInstance must implement

RadiationDetectorDSP.rdfilt!(output, fi::SomeFilterInstance, input)

RadiationDetectorDSP.flt_output_smpltype(fi::SomeFilterInstance)

RadiationDetectorDSP.flt_input_smpltype(fi::SomeFilterInstance)

RadiationDetectorDSP.flt_output_length(fi::SomeFilterInstance)

RadiationDetectorDSP.flt_input_length(fi::SomeFilterInstance)

RadiationDetectorDSP.flt_output_time_axis(fi::SomeFilterInstance, time::AbstractVector{<:RealQuantity})

Invertible filter instances should implement

InverseFunctions.inverse(fi::SomeFilterInstance)

Default methods are implemented for

RadiationDetectorDSP.rdfilt(fi::AbstractRadSigFilterInstance, x::AbstractSamples)

RadiationDetectorDSP.rdfilt(fi::AbstractRadSigFilterInstance, wf::RDWaveform)

RadiationDetectorDSP.bc_rdfilt(fi::AbstractRadSigFilterInstance, inputs)

The default methods that operate on RadiationDetectorSignals.RDWaveforms require RadiationDetectorDSP.flt_output_time_axis.

RadiationDetectorDSP.AbstractSamples — Type

const AbstractSamples{T<:RealQuantity} = AbstractVector{T}

A vector of signal samples.

RadiationDetectorDSP.ArrayOfSimilarSamples — Type

const ArrayOfSimilarSamples{T<:RealQuantity} = ArrayOfSimilarVectors{T}

An array of similar sample vectors.

RadiationDetectorDSP.BiquadFilter — Type

struct BiquadFilter{T<:RealQuantity} <: AbstractRadIIRFilter

A biquad filter.

Constructors:

BiquadFilter(fields...)

Fields:

b_012::Tuple{T, T, T} where T<:(Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real): Coefficients b0 to b2

a_12::Tuple{T, T} where T<:(Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real): Coefficients a1 to a2, a_0 equals one implicitly

RadiationDetectorDSP.CRFilter — Type

struct CRFilter <: AbstractRadIIRFilter

A first-order CR highpass filter.

The inverse filter is InvCRFilter, this is typically stable even in the presence of additional noise.

Constructors:

CRFilter(fields...)

Fields:

cr::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: CR time constant

RadiationDetectorDSP.ConvolutionFilter — Type

struct ConvolutionFilter{T<:RealQuantity} <: AbstractRadFIRFilter

A FIR filter defined by it's filter taps, applied via convolution with the input signal.

Constructors:

ConvolutionFilter(fields...)

Fields:

method::RadiationDetectorDSP.ConvolutionMethod: Convolution method

coeffs::AbstractVector{T} where T<:(Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real): Filter taps

offset::Int64: Time axis offset

RadiationDetectorDSP.ConvolutionMethod — Type

abstract type ConvolutionMethod

Indended as a type parameter to designate the behavior of a filter as linear or nonlinear.

Subtypes are DirectConvolution and FFTConvolution.

RadiationDetectorDSP.DNIMethod — Type

abstract type DNIMethod

Abstract type for denoising and interpolation methods.

RadiationDetectorDSP.DifferentiatorFilter — Type

struct DifferentiatorFilter <: AbstractRadIIRFilter

An integrator filter. It's inverse is IntegratorFilter.

Constructors:

DifferentiatorFilter(fields...)

Fields:

gain::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Filter gain

RadiationDetectorDSP.DirectConvolution — Type

DirectConvolution() isa ConvolutionMethod

Compute filter convolutions directly, without FFT.

RadiationDetectorDSP.FFTConvolution — Type

FFTConvolution() isa ConvolutionMethod

Compute filter convolutions via FFT.

RadiationDetectorDSP.FilteringType — Type

abstract type FilteringType

Indended as a type parameter to designate the behavior of a filter as linear or nonlinear.

Subtypes are LinearFiltering and NonlinearFiltering.

RadiationDetectorDSP.FirstOrderIIR — Type

struct FirstOrderIIR{T<:RealQuantity} <: AbstractRadIIRFilter

A biquad filter.

Constructors:

FirstOrderIIR(fields...)

Fields:

b_01::Tuple{T, T} where T<:(Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real): Coefficients b0 to b1

a_1::Tuple{T} where T<:(Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real): Coefficient a1, a0 equals one implicitly

RadiationDetectorDSP.IntegratorCRFilter — Type

struct IntegratorCRFilter <: AbstractRadIIRFilter

A modified CR-filter. The filter has an inverse.

Constructors:

IntegratorCRFilter(fields...)

Fields:

gain::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Filter gain

cr::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: CR time constant

RadiationDetectorDSP.IntegratorFilter — Type

struct IntegratorFilter <: AbstractRadIIRFilter

An integrator filter. It's inverse is DifferentiatorFilter.

Constructors:

IntegratorFilter(fields...)

Fields:

gain::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Filter gain

RadiationDetectorDSP.IntegratorModCRFilter — Type

struct IntegratorModCRFilter <: AbstractRadIIRFilter

A modified CR-filter. The filter has an inverse.

Constructors:

IntegratorModCRFilter(fields...)

Fields:

gain::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Filter gain

cr::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: CR time constant

RadiationDetectorDSP.InvCRFilter — Type

struct InvCRFilter <: AbstractRadIIRFilter

Inverse of CRFilter.

Constructors:

InvCRFilter(fields...)

Fields:

cr::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: CR time constant

RadiationDetectorDSP.InvModCRFilter — Type

struct InvModCRFilter <: AbstractRadIIRFilter

Inverse of ModCRFilter.

Constructors:

InvModCRFilter(fields...)

Fields:

cr::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: CR time constant

RadiationDetectorDSP.InvRCFilter — Type

struct InvRCFilter <: AbstractRadIIRFilter

Inverse of RCFilter.

Constructors:

InvRCFilter(fields...)

Fields:

rc::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: RC time constant

RadiationDetectorDSP.LinearFiltering — Type

abstract type LinearFiltering <: FilteringType

When used as a type parameter value, marks linear behavior of a filter.

RadiationDetectorDSP.ModCRFilter — Type

struct ModCRFilter <: AbstractRadIIRFilter

A first-order CR highpass filter, modified for full-amplitude step-signal response.

The resonse of the standard digital CRFilter will not recover the full amplitude of a digital step stignal since a step from one sample to the still has a finite rise time. This version of a CR filter compensates for this loss in amplitude, so it effectively treats a step as having

The inverse filter is InvModCRFilter, this is typically stable even in the presence of additional noise.

Constructors:

ModCRFilter(fields...)

Fields:

cr::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: CR time constant

RadiationDetectorDSP.NonlinearFiltering — Type

abstract type NonlinearFiltering <: FilteringType

When used as a type parameter value, marks non linear behavior of a filter.

RadiationDetectorDSP.PolynomialDNI — Type

struct PolynomialDNI <: DNIMethod

Polynomial denoising and interpolation method.

Operates in a similar way as a Savitzky-Golay filter, but interpolates as well.

Constructors:

PolynomialDNI(; fields...)

Fields:

degree::Int64: polynomial degree Default: (2, "length")

length::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real

RadiationDetectorDSP.RCFilter — Type

struct RCFilter <: AbstractRadII>RFilter

A first-order RC lowpass filter.

The inverse filter is InvCRFilter, but note that this is unstable in the presence of additional noise and so will typically not be useful to deconvolve signals in practical applications.

Constructors:

RCFilter(fields...)

Fields:

rc::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: RC time constant

RadiationDetectorDSP.SamplingInfo — Type

struct SamplingInfo{T<:RealQuantity,A<:AbstractVector{<:RealQuantity}}

Holds sampling information.

The numerical type of an individual sample is T, the (time) axis is given by the axis field.

Constructors:

SamplingInfo{T,A}(axis)

SamplingInfo{T}(axis)

Fields:

axis::AbstractVector{<:Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real}

RadiationDetectorDSP.SavitzkyGolayFilter — Type

struct SavitzkyGolayFilter <: AbstractRadFIRFilter

A Savitzky-Golay filter.

Constructors:

SavitzkyGolayFilter(; fields...)

Fields:

length::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: filter length

degree::Int64: Polynomial defgree

derivative::Int64: n-th derivative (0 for no derivative)

RadiationDetectorDSP.SignalEstimator — Type

struct SignalEstimator <: Function

Estimates a signal at a given position x.

Usage:

(f::SamplesOrWaveform)(input::RDWaveform, x::RealQuantity)

Constructors:

SignalEstimator(; fields...)

Fields:

dni::DNIMethod: denoising and interpolation method

RadiationDetectorDSP.SimpleCSAFilter — Type

struct SimpleCSAFilter <: AbstractRadIIRFilter

Simulates the current-signal response of a charge-sensitive preamplifier with resistive reset, the output is a charge signal.

It is equivalent to the composition

CRFilter(cr = tau_decay) ∘
Integrator(gain = gain) ∘
RCFilter(rc = tau_rise)

and maps to a single BiquadFilter.

This filter has an inverse, but the inverse is very unstable in the presence of additional noise if tau_rise is not zero (since the inverse of an RC-filter is unstable under noise). Even if tau_rise is zero the inverse will still amplify noise (since it differentiates), so it should be used very carefully when deconvolving signals in practical applications.

Constructors:

SimpleCSAFilter(fields...)

Fields:

tau_rise::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Rise time constant

tau_decay::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Decay time constant

gain::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Gain

RadiationDetectorDSP.TrapezoidalChargeFilter — Type

struct TrapezoidalChargeFilter <: AbstractRadNonlinearFilter

Filter that responds to a step signal with a trapezoidal pulse.

The filter is equivalent to two moving averages separated by a gap.

Constructors:

TrapezoidalChargeFilter(; fields...)

Fields:

avgtime::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: pre-rise averaging time

gaptime::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: gap time

avgtime2::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: post-rise averaging time

A sharp step on the input will result in a trapezoid with rise time and fall time avgtime and a flat top of length gaptime.

RadiationDetectorDSP.TruncateFilter — Type

struct TruncateFilter <: AbstractRadSigFilter{LinearFiltering}

Filter that truncates the input signal.

Constructors:

TruncateFilter(; fields...)

Fields:

interval::IntervalSets.ClosedInterval{<:Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real}: interval to keep

RadiationDetectorDSP.ZACChargeFilter — Type

struct ZACChargeFilter <: AbstractRadFIRFilter

Zero area cusp (ZAC) filter.

For the definition the filter and a discussion of the filter properties, see "Improvement of the energy resolution via an optimized digital signal processing in GERDA Phase I", Eur. Phys. J. C 75, 255 (2015).

Constructors:

ZACChargeFilter(; fields...)

Fields:

tau::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: equivalent of shaping time (τₛ) Default: (20, "length of flat top (FT)")

toplen::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: Default: 10

length::Union{Unitful.Quantity{<:var"#s1"}, var"#s1"} where var"#s1"<:Real: total length of the filter (L) Default: 100

RadiationDetectorDSP.MaybeWithUnits — Type

const MaybeWithUnits{T<:Number} = Union{T,Quantity{<:T}}

A numerical value with or without units

RadiationDetectorDSP.RealQuantity — Type

const RealQuantity = MaybeWithUnits{<:Real}

A real value with or without units.

RadiationDetectorDSP.SamplesOrWaveform — Type

const RadiationDetectorDSP.SamplesOrWaveform{T<:RealQuantity} = Union{AbstractSamples{T},RDWaveform{<:Any,T}}

A vector of signal samples or a waveform.

RadiationDetectorDSP.add_rect_pulse! — Function

add_rect_pulse!(samples::AbstractSamples, start::Integer, pulselen::Integer, amplitude::Real = 1.0)

Add a rectangular pulse to samples.

RadiationDetectorDSP.bc_rdfilt — Function

bc_rdfilt(flt::AbstractRadSigFilter, input)
bc_rdfilt(fi::AbstractRadSigFilterInstance, input)

Broadcast filter instance fi over signals input, return the filtered signals.

RadiationDetectorDSP.bc_rdfilt! — Function

bc_rdfilt!(outputs, fi::AbstractRadSigFilterInstance, inputs)

Broadcast filter flt or filter instance fi over signals inputs, storing the results in outputs.

inputs and outputs must be of type AbstractVector{<:AbstractSamples}.

Returns outputs.

RadiationDetectorDSP.charge_trapflt! — Method

charge_trapflt!(samples::AbstractVector{<:RealOrSIMD{<:AbstractFloat}}, navg::Integer, ngap::Integer)

Apply a trapezoidal FIR filter to a charge signal in samples.

RadiationDetectorDSP.cr_filter — Method

cr_filter(CR::Real)

Return a DSP.jl-compatible CR-filter.

RadiationDetectorDSP.create_zac_filter — Method

create_zac_filter(Nₜ, FTₜ, τₜ, Δt)

create the zac filter according to Eur. Phys. J. C (2015) 75:255], where Nₜ is the length of the filter, FTₜ the length of the flat top, τₜ the the exponential decay factor and Δt the sampling time

RadiationDetectorDSP.crmod_filter — Method

crmod_filter(CR::Real)

Return a DSP.jl-compatible modified CR-filter.

RadiationDetectorDSP.dfilt! — Function

rdfilt!(output, fi::AbstractRadSigFilterInstance, input)

Apply filter flt or filter instance fi to signal input and store the filtered signal in output. Return output.

RadiationDetectorDSP.differentiator_filter — Method

differentiator_filter(gain::Real)

Return a DSP.jl-compatible differentiator filter.

RadiationDetectorDSP.elsmplinfo — Function

smplinfo(smpls::AbstractSamples)::SamplingInfo
smplinfo(wf::RDWaveform{T,U})::RDWaveform

Get sampling information an array of vectors of samples, resp. an array of waveform. All elements must have equal samling information.

RadiationDetectorDSP.flt_input_length — Function

RadiationDetectorDSP.flt_input_length(fi::AbstractRadSigFilterInstance)::Integer

Get the output signal length of a filter instance fi.

Must be implemented for all subtypes of AbstractRadSigFilterInstance.

RadiationDetectorDSP.flt_input_smpltype — Function

RadiationDetectorDSP.flt_input_smpltype(fi::AbstractRadSigFilterInstance)

Get the input sample type of a filter instance fi.

Must be implemented for all subtypes of AbstractRadSigFilterInstance.

RadiationDetectorDSP.flt_output_length — Function

RadiationDetectorDSP.flt_output_length(fi::SomeFilterInstance)::Integer

Get the output signal length of filter instance fi.

Must be implemented for all subtypes of AbstractRadSigFilterInstance.

RadiationDetectorDSP.flt_output_smpltype — Function

RadiationDetectorDSP.flt_output_smpltype(fi::AbstractRadSigFilterInstance)

Get the output sample type for

a filter flt given an input sample type input_smpltype

a filter instance fi

Must be implemented for all subtypes of AbstractRadSigFilter.

RadiationDetectorDSP.flt_output_time_axis — Function

RadiationDetectorDSP.flt_output_time_axis(fi::SomeFilterInstance, time::AbstractVector{<:RealQuantity})::AbstractVector{<:RealQuantity}

Get the output time axis of a filter instance fi, given an input time axis time.

Must be implemented for subtypes of AbstractRadSigFilter and AbstractRadSigFilterInstance only if the filter's output time axis can be computed directly from the input time axis.

RadiationDetectorDSP.fltinstance — Function

fltinstance(flt::AbstractRadSigFilter, si::SamplingInfo)::AbstractRadSigFilterInstance

Create a filter instance of the filter flt, specialized for the given input, resp. input characteristics.

RadiationDetectorDSP.gen_rect_pulse — Function

gen_rect_pulse(tracelen::Integer, start::Integer, pulselen::Integer, amplitude::Real = 1.0)

Generate a rectangular pulse.

RadiationDetectorDSP.integrator_cr_filter — Method

integrator_cr_filter(gain::Real, CR::Real)

Return a DSP.jl-compatible integrator plus CR filter.

RadiationDetectorDSP.integrator_crmod_filter — Method

integrator_crmod_filter(gain::Real, CR::Real)

Return a DSP.jl-compatible integrator plus modified CR filter.

RadiationDetectorDSP.integrator_filter — Method

integrator_filter(gain::Real)

Return a DSP.jl-compatible integrator filter.

RadiationDetectorDSP.inv_cr_filter — Method

inv_cr_filter(CR::Real)

Return a DSP.jl-compatible inverse CR-filter.

RadiationDetectorDSP.inv_crmod_filter — Method

inv_crmod_filter(CR::Real)

Return a DSP.jl-compatible inverse modified CR-filter.

RadiationDetectorDSP.inv_rc_filter — Method

inv_rc_filter(RC::Real)

Return a DSP.jl-compatible RC-filter.

RadiationDetectorDSP.rc_filter — Method

rc_filter(RC::Real)

Return a DSP.jl-compatible RC-filter.

RadiationDetectorDSP.rdfilt — Function

rdfilt(fi::AbstractRadSigFilterInstance, input)

Apply filter instance fi to signal input, return the filtered signal.

Returns output.

RadiationDetectorDSP.shift_waveform — Function

shift_waveform(signal::AbstractSamples, a::RealQuantity)
shift_waveform(wf::RDWaveform, a::RealQuantity)

Shifts each sample of a waveform up by a.

RadiationDetectorDSP.simple_csa_response_filter — Function

simplecsaresponsefilter(τrise::Real, τdecay::Real, gain::Real = one(τrise))

Return a DSP.jl-compatible filter that models the response of a typical charge-sensitive amplifier (CSA).

RadiationDetectorDSP.smplinfo — Function

smplinfo(smpls::AbstractSamples)::SamplingInfo
smplinfo(wf::RDWaveform{T,U})::RDWaveform

Get sampling information from a vector of samples, resp. a waveform.

RadiationDetectorDSP.zac_charge_filter_coeffs — Method

zac_charge_filter_coeffs(FT, τₛ, Δt, T)

return a vector representing the zac filter applicaible on a charge signal, where FT is the length of the flat top, τₛ the filter shaping time, Δt the sampling time and T the total length of the filter (see Eur. Phys. J. C (2015) 75:255).

RadiationDetectorDSP.zac_current_filter_integral — Method

zac_current_filter_integral(L, FT, τₛ)

Compute the integral of the unnormalised zac filter using its analytical expression, where L is the length of the polynomial part, FT the length of the flat top part and τₛ the filter shaping time

RadiationDetectorDSP.zac_current_filter_shape — Method

zac_current_filter_shape(t, τₛ, L, FT, A)

value of zac filter evaluated at t for the given set of parameters (see zac_charge_filter_coeffs)